Renaming test files to match their purpose.
